refactor import map from standard html plugin to node resolve plugin intercept
